use a toggle button customization demo
